make jobs app and move the data models + migration there
