+++ app/api/config/settings.py
@@ -71,6 +71,7 @@ INSTALLED_APPS = [
     "revproxy",
     "shared",
     "shared.files",
+    "shared.jobs",
 ]
 
 MIDDLEWARE = [
