+++ app/shared/shared/jobs/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FilesConfig(AppConfig):
+    name = "shared.jobs"
+    label = "jobs"
